refactor: 💡 change structure to be modular monolith
