--- src/applications/application.module.ts
@@ -1,8 +1,0 @@
-import { Global, Module } from '@nestjs/common';
-
-@Global()
-@Module({
-  providers: [],
-  exports: [],
-})
-export class ApplicationModule {}
