--- src/domains/item/index.ts
@@ -1,3 +1,0 @@
-export * from './item.model';
-export * from './item-id';
-export * from './item.attribute';
