--- src/domains/item/item-id.ts
@@ -1,3 +1,0 @@
-import { Brand } from '../../applications/types/brand';
-
-export type ItemId = Brand<string, 'ItemId'>;
