--- src/domains/item/item.attribute.ts
@@ -1,4 +1,0 @@
-import { OnlyAttributes } from '../../applications/types';
-import { IItem } from './item.model';
-
-export type ItemAttributes = OnlyAttributes<IItem>;
