+++ src/shares/applications/types/nonFunction.ts
